+++ .gitignore
@@ -1,0 +1,74 @@
+# Compiled source #
+###################
+*.com
+*.class
+*.dll
+*.exe
+*.o
+*.so
+*.pyc
+build/
+dist/
+doc/build/
+
+# Packages #
+############
+# it's better to unpack these files and commit the raw source
+# git has its own built in compression methods
+*.7z
+*.dmg
+*.gz
+*.iso
+*.jar
+*.rar
+*.tar
+*.zip
+
+# Logs and databases #
+######################
+*.log
+*.sql
+*.sqlite
+logs/*
+
+# OS generated files #
+######################
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.Trashes
+.idea
+.tox
+*.sublime*
+*.egg-info
+Icon?
+ehthumbs.db
+Thumbs.db
+.eggs
+
+# User driven backup files #
+############################
+*.bak
+*.swp
+
+# Generated by pbr while building docs
+######################################
+AUTHORS
+ChangeLog
+
+# Files created by releasenotes build
+releasenotes/build
+
+# Test temp files
+tests/plugins
+tests/playbooks
+tests/common
+tests/*.retry
+
+# Vagrant artifacts
+.vagrant
+
+# Eclipse artifacts
+.project
+.pydevproject
