+++ doc/images/leapfrog_structure_diagram.png
