+++ hacking/rpco_hacking_checks.egg-info/PKG-INFO
@@ -1,0 +1,13 @@
+Metadata-Version: 1.1
+Name: rpco-hacking-checks
+Version: 1.0.0
+Summary: Hacking/Flake8 checks for rpc-openstack
+Home-page: UNKNOWN
+Author: Rackspace Private Cloud
+Author-email: UNKNOWN
+License: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
+Classifier: Development Status :: 5 - Production/Stable
+Classifier: Intended Audience :: Developers
+Provides: rpco_checks
