+++ hacking/rpco_hacking_checks.egg-info/SOURCES.txt
@@ -1,0 +1,7 @@
+rpco_checks.py
+rpco_hacking_checks.egg-info/PKG-INFO
+rpco_hacking_checks.egg-info/SOURCES.txt
+rpco_hacking_checks.egg-info/dependency_links.txt
+rpco_hacking_checks.egg-info/entry_points.txt
+rpco_hacking_checks.egg-info/requires.txt
+rpco_hacking_checks.egg-info/top_level.txt
