+++ pre_redeploy.sh
@@ -101,9 +101,7 @@ pushd ${LEAPFROG_DIR}
     fi
     if [[ ! -f "${UPGRADE_LEAP_MARKER_FOLDER}/rebootstrap-ansible-for-rpc.complete" ]]; then
         pushd ${RPCO_DEFAULT_FOLDER}
-            ansible-galaxy install --force -r ansible-role-requirements.yml
-            sed -i "s|GROUP_VARS_PATH=.*|GROUP_VARS_PATH=\"\${GROUP_VARS_PATH:-${RPCO_DEFAULT_FOLDER}/openstack-ansible/playbooks/inventory/group_vars/:${RPCO_DEFAULT_FOLDER}/group_vars/:/etc/openstack_deploy/group_vars/}\"|" /usr/local/bin/openstack-ansible.rc
-            sed -i "s|HOST_VARS_PATH=.*|HOST_VARS_PATH=\"\${HOST_VARS_PATH:-${RPCO_DEFAULT_FOLDER}/openstack-ansible/playbooks/inventory/host_vars/:${RPCO_DEFAULT_FOLDER}/host_vars/:/etc/openstack_deploy/host_vars/}\"|" /usr/local/bin/openstack-ansible.rc
+            scripts/bootstrap-ansible.sh
             source /usr/local/bin/openstack-ansible.rc
         popd
         log "rebootstrap-ansible-for-rpc" "ok"
