+++ tests/ansible-env.rc
@@ -1,4 +1,3 @@
----
 # Copyright 2017, Rackspace US, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,16 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-- hosts: hosts
-  gather_facts: no
-  tasks:
-    - name: Get list of old containers
-      command: "cat /etc/openstack_deploy/leapfrog_remove_remaining_old_containers"
-      run_once: true
-      delegate_to: localhost
-      register: old_containers
-
-    - name: Remove old MaaS neutron container config files
-      shell: "rm /etc/rackspace-monitoring-agent.conf.d/*{{ item }}*"
-      with_items: old_containers.stdout_lines
-      ignore_errors: true
+export ANSIBLE_FILTER_PLUGINS="/opt/openstack-ansible/playbooks/plugins/filters"
+export ANSIBLE_LOOKUP_PLUGINS="/opt/openstack-ansible/playbooks/plugins/lookups"
+export ANSIBLE_CALLBACK_PLUGINS="/opt/openstack-ansible/playbooks/plugins/callbacks"
+export ANSIBLE_ACTION_PLUGINS="/opt/openstack-ansible/playbooks/plugins/actions"
+export ANSIBLE_GATHERING="smart"
+export ANSIBLE_CACHE_PLUGIN="jsonfile"
+export ANSIBLE_CACHE_PLUGIN_CONNECTION="/etc/openstack_deploy/ansible_facts"
+export ANSIBLE_CACHE_PLUGIN_TIMEOUT="86400"
+export ANSIBLE_HOST_KEY_CHECKING="false"
