+++ tox.ini
@@ -1,0 +1,143 @@
+[tox]
+minversion = 2.0
+skipsdist = True
+envlist = bindep,docs,linters,functional
+sitepackages = True
+
+
+[testenv]
+usedevelop = True
+install_command =
+    pip install -c{env:UPPER_CONSTRAINTS_FILE} {opts} {packages} --isolated
+deps =
+    -r{toxinidir}/test-requirements.txt
+commands =
+    /usr/bin/find . -type f -name "*.pyc" -delete
+passenv = *
+whitelist_externals =
+    bash
+setenv =
+    BINDEP_FILE={toxinidir}/bindep.txt
+    PYTHONUNBUFFERED=1
+    TEST_IDEMPOTENCE=false
+    VIRTUAL_ENV={envdir}
+    WORKING_DIR={toxinidir}
+    # RPC Gate variables
+    IRR_CONTEXT={env:IRR_CONTEXT:master}
+    # bug#1682108
+    PYTHONPATH={envsitepackagesdir}
+    ### OSA specific call back files
+    # Set the checkout to any supported tag, branch, or sha.
+    OSA_TESTS_CHECKOUT=master
+    UPPER_CONSTRAINTS_FILE=https://git.openstack.org/cgit/openstack/requirements/plain/upper-constraints.txt?h={env:OSA_TESTS_CHECKOUT:master}
+    OSA_TEST_DEPS=https://git.openstack.org/cgit/openstack/openstack-ansible-tests/plain/test-ansible-deps.txt?h={env:OSA_TESTS_CHECKOUT:master}
+
+
+[testenv:bindep]
+# Deps is empty on purpose
+deps =
+commands =
+    bash -c "{toxinidir}/run-bindep.sh"
+
+
+[testenv:docs]
+commands =
+    bash -c "rm -rf doc/build"
+    doc8 doc
+    python setup.py build_sphinx
+
+
+[doc8]
+# Settings for doc8:
+extensions = .rst
+
+
+[testenv:releasenotes]
+commands =
+    sphinx-build -a -E -W -d releasenotes/build/doctrees -b html releasenotes/source releasenotes/build/html
+
+
+# environment used by the -infra templated docs job
+[testenv:venv]
+commands =
+    {posargs}
+
+
+[testenv:tests_clone]
+commands =
+    bash -c "if [ ! -d "{toxinidir}/tests/common" ]; then \
+               git clone https://git.openstack.org/openstack/openstack-ansible-tests {toxinidir}/tests/common; \
+               pushd {toxinidir}/tests/common; \
+                 git checkout {env:OSA_TESTS_CHECKOUT:master}; \
+               popd; \
+             fi"
+
+
+[testenv:pep8]
+commands =
+    {[testenv:tests_clone]commands}
+    bash -c "{toxinidir}/tests/common/test-pep8.sh"
+
+
+[flake8]
+# Ignores the following rules due to how ansible modules work in general
+#     F403 'from ansible.module_utils.basic import *' used;
+#          unable to detect undefined names
+ignore=F403,E731
+
+
+[testenv:bashate]
+commands =
+    {[testenv:tests_clone]commands}
+    bash -c "{toxinidir}/tests/common/test-bashate.sh"
+
+
+[testenv:ansible]
+deps =
+    {[testenv]deps}
+    -r{env:OSA_TEST_DEPS}
+
+
+[testenv:ansible-syntax]
+deps =
+    {[testenv:ansible]deps}
+commands =
+    {[testenv:tests_clone]commands}
+    bash -c "{toxinidir}/tests/common/test-ansible-syntax.sh"
+
+
+[testenv:ansible-lint]
+deps =
+    {[testenv:ansible]deps}
+commands =
+    {[testenv:tests_clone]commands}
+    bash -c "{toxinidir}/tests/common/test-ansible-lint.sh"
+
+
+[testenv:linters]
+deps =
+    {[testenv:ansible]deps}
+commands =
+    {[testenv:tests_clone]commands}
+    {[testenv:pep8]commands}
+    {[testenv:bashate]commands}
+    {[testenv:ansible-lint]commands}
+    {[testenv:ansible-syntax]commands}
+    {[testenv:docs]commands}
+
+
+[testenv:leapfrog]
+commands =
+    bash -c "{toxinidir}/tests/test-leapfrog.sh"
+
+
+[testenv:functional]
+install_command =
+    pip install {opts} {packages} --isolated --upgrade --force-reinstall
+deps =
+    -r{toxinidir}/test-requirements.txt
+commands =
+    {[testenv:bindep]commands}
+    {[testenv:tests_clone]commands}
+    bash -c "{toxinidir}/tests/aio-create.sh"
+    {[testenv:leapfrog]commands}
